// src/models/IconName.ts
export enum IconName {
  admin = 'admin',
  arrowDown = 'arrowDown',
  babToken = 'babToken',
  babFemale = 'babFemale',
  babMale = 'babMale',
  babNeutral = 'babNeutral',
  backArrow = 'backArrow',
  bars = 'bars',
  bear = 'bear',
  bellActive = 'bellActive',
  bellDisabled = 'bellDisabled',
  bolt = 'bolt',
  book = 'book',
  bow = 'bow',
  bull = 'bull',
  cancel = 'cancel',
  compound = 'compound',
  controls = 'controls',
  chat = 'chat',
  chartUp = 'chartUp',
  check = 'check',
  chevronDown = 'chevronDown',
  chevronUp = 'chevronUp',
  clipboard = 'clipboard',
  coin = 'coin',
  coinStack = 'coinStack',
  customGarden = 'customGarden',
  creator = 'creator',
  dialogue = 'dialogue',
  discord = 'discord',
  edit = 'edit',
  error = 'error',
  exit = 'exit',
  external = 'external',
  failure = 'failure',
  flame = 'flame',
  garden = 'garden',
  gate = 'gate',
  gear = 'gear',
  heartFull = 'heartFull',
  heartHalf = 'heartHalf',
  identity = 'identity',
  immunefi = 'immunefi',
  link = 'link',
  lock = 'lock',
  lock2 = 'lock2',
  management = 'management',
  merman = 'merman',
  member = 'member',
  members = 'members',
  newBadge = 'newBadge',
  nocode = 'nocode',
  numberDown = 'numberDown',
  numberUp = 'numberUp',
  opensea = 'opensea',
  permission = 'permission',
  pending = 'pending',
  plus = 'plus',
  question = 'question',
  rightArrow = 'rightArrow',
  rocket = 'rocket',
  schedule = 'schedule',
  shield = 'shield',
  starCrashing = 'starCrashing',
  starShooting = 'starShooting',
  steward = 'steward',
  stewardPrivate = 'stewardPrivate',
  stewardPublic = 'stewardPublic',
  stop = 'stop',
  strategist = 'strategist',
  strategistPrivate = 'strategistPrivate',
  strategistPublic = 'strategistPublic',
  success = 'success',
  switch = 'switch',
  trash = 'trash',
  transfer = 'transfer',
  upload = 'upload',
  video = 'video',
  warning = 'warning',
  xLarge = 'xLarge',
}